move to sway, yay no tearing \o/

+++ update.sh
@@ -17,15 +17,23 @@ cp ~/.config/nvim/treesitter.vim ./.config/nvim/
 cp ~/.config/nvim/vsnip.vim ./.config/nvim/
 cp -R ~/.config/nvim/after/. ./.config/nvim/after
 
-# snippets
-cp ~/.vsnip/*.json ./.vsnip/
-
 # kitty
-cp ~/.config/kitty/kitty.conf ./.config/kitty/ 
+if [ -e "$HOME/.config/kitty/kitty.conf" ]; then
+  cp ~/.config/kitty/kitty.conf ./.config/kitty/ 
+fi
 
 # tmux
 cp ~/.config/tmux/tmux.conf ./.config/tmux/tmux.conf
 
 # i3 & polybar
-cp -R ~/.config/i3/. ./.config/i3
-cp -R ~/.config/polybar/. ./.config/polybar
+if [ -e "$HOME/.config/i3" ]; then
+	cp -R ~/.config/i3/. ./.config/i3
+fi
+if [ -e "$HOME/.config/polybar" ]; then
+	cp -R ~/.config/polybar/. ./.config/polybar
+fi
+
+# sway & waybar & rofi
+cp -R ~/.config/sway ./.config/sway
+cp -R ~/.config/waybar ./.config/waybar
+cp -R ~/.config/rofi ./.config/rofi
